[Product] Implement category drop down list & merge create and edit into upsert
- Category drop down
- Product View model
- Create & Edit view merge
- Fix redirect bug

+++ BulkyBookWeb/Areas/Admin/Controllers/CategoryController.cs
@@ -108,6 +108,10 @@ public class CategoryController(
         var obj = dbFactory
             .CategoryRepository
             .Get(u => u.Id == id);
+        if (obj == null)
+        {
+            return;
+        }
         dbFactory.CategoryRepository.Delete(obj);
         dbFactory.Save();
     }
